backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
sqlalchemy==2.0.43
asyncpg==0.30.0
geoalchemy2==0.17.1
alembic==1.16.4
python-jose[cryptography]==3.5.0
httpx==0.28.1
redis==6.4.0
orjson==3.11.3
python-multipart==0.0.20
